# api/openapi.yaml
openapi: 3.1.0
info:
  title: BFC VPN API
  version: 1.0.0
  description: |
    BFC VPN Management API - Vietnamese messages for all error responses.

    ## Authentication
    Most endpoints require JWT Bearer token in Authorization header.

    ## Error Format
    All errors follow RFC 7807 Problem Details format.
  contact:
    name: BFC VPN Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8081
    description: Development
  - url: https://api.bfc-vpn.com
    description: Production
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: authLogin
      tags: [Auth]
      summary: User login
      description: Authenticate user with email/password. Returns tokens or requires MFA.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful or MFA required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 12
          maxLength: 128
    LoginResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [success, mfa_required]
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        mfa_token:
          type: string
          description: Temporary token for MFA verification
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Vietnamese error message for end users
        instance:
          type: string
          format: uri
          description: URI reference to this specific occurrence
        action:
          type: string
          description: Vietnamese suggested action for users
        errors:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors (Vietnamese)
        request_id:
          type: string
          description: Request ID for support reference